backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
google-genai==0.3.0
google-cloud-storage==2.14.0
google-cloud-logging==3.9.0
pydantic==2.6.0
pydantic-settings==2.1.0
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.27.0
